# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BLOG API",
    "description": "Description",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3300/"
    }
  ],
  "paths": {
...
    "/category/list": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get Category list",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {}
      }
    },
    "/category/": {
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Create Category",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getCategoryCreateObject"
              },
              "example": {
                "name": "category Name"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "Update Category",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getCategoryUpdateObject"
              },
              "example": {
                "_id": "...(categoryId)",
                "name": "example category"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Delete Category",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getCategoryDelObject"
              },
              "example": {
                "_id": "...(postId)"
              }
            }
          }
        }
      }
    },
...
    "/post/list": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get post list",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "Create post list",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getPostCreateArrayObject"
              },
              "example": [
                {
                  "author": "...(authorId)",
                  "title": "test title",
                  "authorName": "testUser",
                  "summary": "example summary",
                  "text": "example text",
                  "relatedPosts": []
                },
                {
                  "author": "...(authorId)",
                  "authorName": "testUser",
                  "title": "test title2",
                  "summary": "example summary2",
                  "text": "example text2",
                  "relatedPosts": []
                }
              ]
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Post"
        ],
        "summary": "Delete post list",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getPostCreateArrayObject"
              },
              "example": [
                {
                  "_id": "...(postId)"
                },
                {
                  "_id": "...(postId)"
                }
              ]
            }
          }
        }
      }
    },
    "/post/": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Get post",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/_id"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Post"
        ],
        "summary": "Create post",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getPostCreateObject"
              },
              "example": {
                "author": "...(authorId)",
                "authorName": "testUser",
                "summary": "example summary",
                "title": "example title",
                "text": "example text",
                "parent": "",
                "relatedPosts": [],
                "tags": []
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Post"
        ],
        "summary": "Update post",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getPostCreateArrayObject"
              },
              "example": {
                "_id": "...(postId)",
                "summary": "example summary",
                "text": "example text",
                "relatedPosts": []
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Post"
        ],
        "summary": "Delete post",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/getPostCreateArrayObject"
              },
              "example": {
                "_id": "...(postId)"
              }
            }
          }
        }
      }
    },
    "/post/list/search": {
      "get": {
        "tags": [
          "Post"
        ],
        "summary": "Search post list",
        "description": "",
        "parameters": [
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/word"
          }
        ],
        "responses": {}
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "getPostCreateArrayObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "array"
          },
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "object"
              },
              "properties": {
                "type": "object",
                "properties": {
                  "author": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "pattern": {
                        "type": "string",
                        "example": "^[a-fA-F0-9]{24}$"
                      },
                      "title": {
                        "type": "string",
                        "example": "작성자"
                      }
                    }
                  },
                  "authorName": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "pattern": {
                        "type": "string",
                        "example": "^([a-zA-Z]|[가-힣]|[0-9])+$"
                      },
                      "minLength": {
                        "type": "number",
                        "example": 3
                      },
                      "maxLength": {
                        "type": "number",
                        "example": 20
                      },
                      "title": {
                        "type": "string",
                        "example": "작성자 이름"
                      }
                    }
                  },
                  "summary": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "minLength": {
                        "type": "number",
                        "example": 3
                      },
                      "maxLength": {
                        "type": "number",
                        "example": 100
                      },
                      "title": {
                        "type": "string",
                        "example": "요약"
                      }
                    }
                  },
                  "text": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "minLength": {
                        "type": "number",
                        "example": 3
                      },
                      "maxLength": {
                        "type": "number",
                        "example": 10000
                      },
                      "title": {
                        "type": "string",
                        "example": "내용"
                      }
                    }
                  },
                  "parent": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "pattern": {
                        "type": "string",
                        "example": "^[a-fA-F0-9]{24}$"
                      },
                      "title": {
                        "type": "string",
                        "example": "부모 게시물"
                      }
                    }
                  },
                  "relatedPosts": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "array"
                      },
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "pattern": {
                            "type": "string",
                            "example": "^[a-fA-F0-9]{24}$"
                          }
                        }
                      },
                      "title": {
                        "type": "string",
                        "example": "관련된 게시물"
                      }
                    }
                  },
                  "tags": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "array"
                      },
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "pattern": {
                            "type": "string",
                            "example": "^[a-fA-F0-9]{24}$"
                          }
                        }
                      },
                      "title": {
                        "type": "string",
                        "example": "태그"
                      }
                    }
                  },
                  "category": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "pattern": {
                        "type": "string",
                        "example": "^[a-fA-F0-9]{24}$"
                      },
                      "title": {
                        "type": "string",
                        "example": "카테고리"
                      }
                    }
                  }
                }
              },
              "required": {
                "type": "array",
                "example": [
                  "author",
                  "authorName",
                  "summary",
                  "text"
                ],
                "items": {
                  "type": "string"
                }
              },
              "additionalProperties": {
                "type": "boolean",
                "example": false
              },
              "title": {
                "type": "string",
                "example": "게시물 작성"
              }
            }
          },
          "minItems": {
            "type": "number",
            "example": 1
          },
          "maxItems": {
            "type": "number",
            "example": 1000
          },
          "title": {
            "type": "string",
            "example": "게시물 작성 리스트"
          }
        }
      },
...
      "getCategoryCreateObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 40
                  },
                  "title": {
                    "type": "string",
                    "example": "이름"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "name"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "카테고리 생성"
          }
        }
      },
      "getCategoryUpdateObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "카테고리 아이디"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "minLength": {
                    "type": "number",
                    "example": 1
                  },
                  "maxLength": {
                    "type": "number",
                    "example": 40
                  },
                  "title": {
                    "type": "string",
                    "example": "이름"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "_id",
              "name"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "카테고리 갱신"
          }
        }
      },
      "getCategoryDelObject": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "pattern": {
                    "type": "string",
                    "example": "^[a-fA-F0-9]{24}$"
                  },
                  "title": {
                    "type": "string",
                    "example": "카테고리 아이디"
                  }
                }
              }
            }
          },
          "required": {
            "type": "array",
            "example": [
              "_id"
            ],
            "items": {
              "type": "string"
            }
          },
          "additionalProperties": {
            "type": "boolean",
            "example": false
          },
          "title": {
            "type": "string",
            "example": "카테고리 삭제"
          }
        }
      }
    },
    "parameters": {
...
      "limit": {
        "in": "query",
        "name": "limit",
        "required": true,
        "description": "가져올 리스트 갯수",
        "schema": {
          "type": "number"
        }
      },
      "offset": {
        "in": "query",
        "name": "offset",
        "required": true,
        "description": "리스트 스킵 갯수",
        "schema": {
          "type": "number"
        }
      },
      "_id": {
        "in": "query",
        "name": "_id",
        "required": true,
        "description": "카테고리 아이디",
        "schema": {
          "type": "string"
        }
      },
      "word": {
        "in": "query",
        "name": "word",
        "required": false,
        "description": "검색어",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}